requirements.txt
streamlit
opencv-python-headless
mediapipe
scikit-learn
numpy
pillow
